Moved youtube commands to features
