+++ commands/management.py
@@ -1,7 +1,6 @@
 import data
 import features
 import logging
-import pytube
 from .helper import handle_command_exception
 from bot import client
 from nextcord import Interaction, Permissions, SlashOption
@@ -362,126 +361,3 @@ async def settings(
                 await interaction_response.edit("Operation failed.")
     except:
         await handle_command_exception("settings", interaction, interaction_response)
-
-
-@client.slash_command(
-    name="youtube",
-    description="Send new youtube videos in a channel.",
-    default_member_permissions=Permissions(administrator=True),
-    dm_permission=False,
-)
-async def youtube_notification_set(
-    interaction: Interaction,
-    link: str = SlashOption(
-        required=True, description="A video link from the target youtube channel"
-    ),
-    channel_id: str = SlashOption(
-        required=False,
-        description="Target Discord channel id to publish new youtube videos.",
-    ),
-    custom_message: str = SlashOption(
-        required=False,
-        description="Use '\\n' for new line. Leave empty to use the default message.",
-    ),
-):
-    try:
-        _logger.info(
-            "commands/management: Command 'youtube' was called by "
-            + f"'{interaction.user.name}' ({interaction.user.id}) "
-            + f"in '{interaction.guild.name}' ({interaction.guild_id}) "
-            + f"args: link:{link} channel_id:{channel_id}"
-        )
-        interaction_response = await interaction.send("Please wait...", ephemeral=True)
-
-        if not features.youtube_notify.is_active():
-            await interaction_response.edit(
-                "Sorry! This feature is unavailable at the moment...",
-            )
-            return
-
-        if not channel_id:
-            channel_id = interaction.channel_id
-
-        channel_id = int(channel_id)
-        channel = client.get_channel(channel_id)
-        if interaction.guild_id != channel.guild.id:
-            _logger.debug(
-                "commands/management: Invalid discord channel "
-                + "or the specified channel belongs to another guild "
-                + f"channel_id: {channel_id}"
-            )
-            await interaction_response.edit(
-                "Invalid channel ID.",
-            )
-            return
-
-        video = pytube.YouTube(link)
-        last_channel_video = features.youtube_notify.get_last_video_of_youtube_channel(
-            video.channel_id
-        )
-
-        result = data.yt_notif_rule_add(
-            interaction.guild_id,
-            video.channel_id,
-            video.author,
-            channel_id,
-            last_channel_video["id"],
-            custom_message,
-        )
-        if result == video.channel_id or result == "Updated.":
-            _logger.info(
-                f"commands/management: {video.author}'s new youtube videos "
-                + f"will be posted on '{channel.name}' ({channel.id})"
-            )
-            await interaction_response.edit(
-                f"Done. **{video.author}** new videos will be posted on **{channel.name}**.",
-            )
-        else:
-            await interaction_response.edit(str(result))
-    except:
-        await handle_command_exception("youtube", interaction, interaction_response)
-
-
-@client.slash_command(
-    name="youtube_remove",
-    description="Remove a previously set notification rule.",
-    default_member_permissions=Permissions(administrator=True),
-    dm_permission=False,
-)
-async def youtube_notification_remove(
-    interaction: Interaction,
-    link: str = SlashOption(
-        required=True, description="A video link from the target youtube channel"
-    ),
-):
-    try:
-        _logger.info(
-            "commands/management: Command 'youtube_remove' was called by "
-            + f"'{interaction.user.name}' ({interaction.user.id}) "
-            + f"in '{interaction.guild.name}' ({interaction.guild_id}) args: link:{link}"
-        )
-        interaction_response = await interaction.send("Please wait...", ephemeral=True)
-
-        if not features.youtube_notify.is_active():
-            await interaction_response.edit(
-                "Sorry! This feature is unavailable at the moment...",
-            )
-            return
-
-        video = pytube.YouTube(link)
-
-        result = data.yt_notif_rule_remove(interaction.guild_id, video.channel_id)
-        if result == video.channel_id:
-            _logger.info(
-                f"commands/management: {video.author}'s new youtube videos will no longer "
-                + f"be posted on '{interaction.guild.name}' ({interaction.guild_id})"
-            )
-            await interaction_response.edit(
-                f"You will no longer receive new videos from **{video.author}**.",
-            )
-        else:
-            await interaction_response.edit(str(result))
-    except:
-        await handle_command_exception(
-            "youtube_remove", interaction, interaction_response
-        )
