+++ features/management/admin/__init__.py
@@ -1,3 +1,4 @@
 from . import config_report
 from . import set_role
+from . import youtube
 from .feature import *
