+++ features/management/admin/feature.py
@@ -1,6 +1,7 @@
 import logging
 from . import config_report
 from . import set_role
+from . import youtube
 
 _logger = logging.getLogger("main")
 
@@ -20,3 +21,4 @@ def activate():
     _logger.debug("features: Feature has been activated: 'management/admin'")
     config_report.activate()
     set_role.activate()
+    youtube.activate()
